separate game logic
