+++ lib/player.rb
@@ -1,0 +1,9 @@
+class Player
+    attr_reader :name, :token
+  
+    def initialize(name, token)
+      @name = name
+      @token = token
+    end
+  end
+
